+++ packages/types/src/condition.ts
@@ -19,8 +19,19 @@ type AndCondition<TDescription extends Description> = {
   and: Condition<TDescription>[]
 }
 
+type ExistsCondition<TDescription extends Description> = {
+  operator: 'exists'
+  term1: PropertiesWithId<TDescription>
+}
+
+type NotCondition<TDescription extends Description> = {
+  not: Condition<TDescription>
+}
+
 export type Condition<TDescription extends Description> =
   | FinalCondition<TDescription>
+  | ExistsCondition<TDescription>
+  | NotCondition<TDescription>
   | OrCondition<TDescription>
   | AndCondition<TDescription>